+++ imow/common/consts.py
@@ -1,0 +1,2 @@
+IMOW_OAUTH_URI = "https://oauth2.imow.stihl.com"
+IMOW_API_URI = "https://api.imow.stihl.com"