+++ tests/test_unit_imow.py
@@ -2,8 +2,6 @@
 import unittest
 
 from imow.api import IMowApi
-from imow.common.mower import Mower, MowerState
-
 TOKEN = 'xx508xx63817x752xx74004x30705xx92x58349x5x78f5xx34xxxxx51'
 MOWER_NAME = "MyMower"
 MOWER_ACTION_ID = "0000000123456789"
@@ -23,16 +21,3 @@ class TestIMowApiUnit(unittest.TestCase):
     def test_auth_without_anything(self):
         with self.assertRaises(LookupError, msg="Expected API to raise when instancing without credentials"):
             IMowApi()
-
-    def test_unknown_mower_state(self):
-        result = Mower(dict())
-        result.status["mainState"] = 123
-        self.assertIsInstance(result.get_current_state(), int, msg="Expected Mower class returned int(123)")
-        self.assertEqual(result.get_current_state(), 123, msg="Expected Mower class returned int(123)")
-
-    def test_known_mower_state(self):
-        result = Mower(dict())
-        result.status["mainState"] = 7
-        self.assertIsInstance(result.get_current_state(), MowerState, msg="Expected Mower class returned int(123)")
-        self.assertEqual(result.get_current_state(), MowerState.INSIDE_DOCK_CHARGING,
-                         msg="Expected Mower class returned MowerState.INSIDE_DOCK_CHARGING")